Move PostAdapter to the right package
